+++ doc/source/changelog.rst
@@ -17,9 +17,10 @@ master: 1.7.2
 * Capture Cling error messages from cppdef and include in the Python exception
 * Add a cppexec method to evalutate statements in Cling's global scope
 * Support initialization of std::array<> from sequences
-* Support C++17 style initialization of STL containers
+* Support C++17 style initialization of common STL containers
 * Allow base classes with no virtual destructor (with warning)
 * Support const by-value returns in Python-side method overrides
+* Support for cross-language multiple inheritance of C++ bases
 
 
 2020-06-06: 1.7.1