+++ doc/source/changelog.rst
@@ -23,6 +23,7 @@ master: 1.7.2
 * Support for cross-language multiple inheritance of C++ bases
 * Allow for pass-by-value of std::unique_ptr through move
 * Reduced dependencies on upstream code
+* Put remaining upstream code in CppyyLegacy namespace
 
 
 2020-06-06: 1.7.1